set up logrus, with optional env vars for syslog
Note that gregord vendors its own copy of the rpc directory, so we have
to make these changes twice. This in turn is because vendoring things
properly at the top level of the gregor project causes conflicts, when
other projects that vendor gregor have different versions of the same
dependencies. This is a nightmare.