improve the filenames in gregor log output
Previously we were wrapping the logger object but not setting
ExtraCalldepth, so all our log lines uselessly referred to the log.go
wrappers. Even after fixing this, the use of "shortfile" was a little
confusing, because it didn't distinguish between "rpc/log.go" and
"daemons/gregord/log.go". Switching to "longfile" helps with this,
though it's somewhat more verbose.